wobbly: Drop dependency on boost, use wobbly::geometry helpers
This change unfortunately needs to touch a lot of code. The boost
dependency was dropped as building against boost on Debian is a
pain - the API keeps changing with each major release and it is
a very heavy package level dependency.

In some cases, we also needed helpers in place to make certain
functionality (such as ostream overloads or MPL asserts) work
correctly. Since we can rely on C++17 support, there's no need
for such hacks and they can be dropped.

https://phabricator.endlessm.com/T23173