wobbly: Fix model target calculation when releasing second anchor
This was caused by two things. First, we were not updating the
target mesh origin when the anchor count *dropped* to 1 (only
when it reached 1) and second, we were using the old targets
to constrain the expected target position.

This commit updates the mesh origin when the anchor count drops
to 1 and it also performs a full integration to work out the
target position, as opposed to relying on an optimisation
which would have been incorrect.