animation: Add bounce animation
This animation simulates a gentle bounce on a sine wave from
the center of the window outwards.

The best way to think of this animation is to think of an attenuating
sine wave squeezed inwards by two bounds that converge on a single point.

In thise case those lines are running from the "initialScale" to 1.0
and the "maximumScale" down to 1.0. We run the sine wave for 2pi * nBounce
iterations (scaling it it to fit within the time range), but the effect
of the naimation is scaled according to where we are on the bunds.

Now, rotate the attenuating sine wave so that it is facing towards you
on the z-axis (OpenGL co-ordinates). This is essentially what the animation
is.