animation: Add magic lamp animation
This animation simulates the "genie" effect on another popular
operating system, appearing to "suck" the window into a destination
rectangle.

The way it works is a little complicated. Basically, we subdivide
the surface into a very high resolution grid and from each point
in the source geometry to each point in the target geometry, we
we have a sigmoid function (S-curve, eg e^x / (1 + ke^px) that each
grid point travels down towards its destination.

When deforming a UV into the pre-computed grid we must do a linear
interpolation along both axes, finding the nearest subdivision source
point on a UV mesh, then interpolating into that mesh component
on the scene-relative internal mesh.

https://phabricator.endlessm.com/T23697